requirements.txt
numpy==1.23.0
pillow==9.2.0
opencv-python==4.5.2.52
onnxruntime==1.11.1
torchvision==0.13.0
streamlit==1.22.0